refact: extrair handlers de trade P2P para module/trade/trade-handlers.mjs
