+++ packs/efeitos-cardigan/LOG
@@ -1,0 +1,3 @@
+2026/06/30-20:33:51.019 4104 Recovering log #1285
+2026/06/30-20:33:51.061 4104 Delete type=0 #1285
+2026/06/30-20:33:51.061 4104 Delete type=3 #1284
