+++ packs/equipamentos-cardigan/LOG
@@ -1,0 +1,3 @@
+2026/06/30-20:33:51.231 4104 Recovering log #1296
+2026/06/30-20:33:51.324 4104 Delete type=0 #1296
+2026/06/30-20:33:51.324 4104 Delete type=3 #1295
