+++ packs/racas-cardigan/LOG
@@ -1,0 +1,3 @@
+2026/06/30-20:33:51.171 3704 Recovering log #2244
+2026/06/30-20:33:51.211 3704 Delete type=0 #2244
+2026/06/30-20:33:51.211 3704 Delete type=3 #2243
