+++ packs/skills-cardigan/LOG
@@ -1,0 +1,3 @@
+2026/06/30-20:33:51.091 2ea4 Recovering log #1313
+2026/06/30-20:33:51.141 2ea4 Delete type=0 #1313
+2026/06/30-20:33:51.141 2ea4 Delete type=3 #1312
